pyproject.toml
[tool.poetry]
name = "paw-patrol"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.6.1"
numpy = "^1.26.4"
pillow = "^10.2.0"
torch = "^2.2.0"
torchvision = "^0.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
